requirements.txt
cloudinary  # use latest and hopw for the best
python-dotenv
pydantic==2.9.0
Sphinx==8.0.2
